typo in /etc/init.d/bootopts.sh

+++ etc/init.d/bootopts.sh
@@ -35,7 +35,7 @@ if ! grep -q "100[0-9]:100" /etc/passwd; then
 	fi
 	
 	# Make sure we have users applications.conf
-	if [ ! -f "/etc/skel/.config/slitaz/applications.conf" -a
+	if [ ! -f "/etc/skel/.config/slitaz/applications.conf" -a \
 	     -f "/etc/slitaz/applications.conf" ]; then
 		mkdir -p /etc/skel/.config/slitaz
 		cp /etc/slitaz/applications.conf /etc/skel/.config/slitaz
